Title: Minor clean up for stricter checks, casting style, and doc types
These improvements touch areas around cleaning up deprecated functions
such as intval(), strval() into casting with (int), (string), etc.

If such improvements is welcomed into the Title class, I can make more
of such but for now, I've just made a few, like 1/20 of these kind of
changes in the file so that, if it's merged and encouraged to continue,
I'll cleanup the file. Or as proposed on this patch, phpcs rules can be
used to handle such cases.

Change-Id: Ib90e880cb124e765379ddad531d4c89289d364de