resources: Remove 'jquery.xmldom' module
Added in r93908 (2011, aa1bc6f742c0), for use by UploadWizard,
but as far as I can tell this never happened, and it's not
in use currently anywhere in Wikimedia Git or elsewhere indexed
by Codesearch.

It's never been updated or otherwise seen maintenance, and seems
obsolete with native browser features nowadays.

Bug: T202154
Change-Id: I6eccf4f488b6e8d7afa5eb81505895d0729cbfea