resources: Remove deprecated 'jquery.hidpi' module
Deprecated in 1.32 and has no further purpose in its current
form as a jQuery plugin for <img srcset>, which we now use
natively without fallback.

The remaining logic for bracketed window.devicePixelRatio is
simple enough to inline as needed without the cruft and overhead
that comes with a centralised approach.

Bug: T202154
Change-Id: I729dfabcbb40a0a794d6b166a584f45a64ac0338