resourceloader: Remove warning for handled localStorage error
Leads to more confusion that actual utility in practice.
The storage being disabled is an entirely supported and expected
scenario, which we handle gracefully.

When wanting to ad-hoc check whether localStorage is enabled,
browsers tend to show that in their interface already (e.g. to
the side of the address bar in Chrome). And it can also be
trivially checked from the console by looking at localStorage
or mw.loader.store.

Bug: T195647
Change-Id: I1edf89e52ec56b4919816878924f096d5b7f7751