resourceloader: Remove wgAvailableSkins from mw.config
Added in 2011 for use in QUnit tests (r80790, r80792).

That was a terrible idea in hindsight, which I removed again in
2015 with 6b758fc982972e, and 0f9e4ca0fb7c.

If a need for it were to arise in the future, we'd export it
within the file module that needs it, or via addJsConfigVars()
on only the pages that need it – not globally.

Bug: T217772
Change-Id: I3885cacaa9b33e6947dbaf26f9b6839e19588603