build: Remove failing HHVM test from Travis CI config
This is the only version of HHVM that MediaWiki still supports,
and while it is passing on WMF's Jenkins, its redundant counterpart
on Travis has been failing for months.

It doesn't appear to be adding value in its current form,
and it's unlikely that after failing so long, that someone would
invest in making it pass before support for HHVM is dropped later
this year (in master, which is the only branch we test in Travis).

It's also delayed feedback by a third (~10min -> 15min), because
it seems to be much slower than the others (unsure why).

Change-Id: I89bd09cb04afde89c8e8ec9a06f69ecb4b094714