resourceloader: Separate context from extraQuery in ClientHtmlTest
The test was dual-purposing the same array both to create
a ResourceLoaderContext object, and to pass further down as
$extraQuery parameter.

This meant that ResourceLoaderClientHtml::makeLoad, and indirectly
ResourceLoader::makeLoaderQuery, did not have any control over
the formatting of the query string, due to it forcing the key/value
pairs regardless (in addition to being redundant).

This is needed in order to omit debug=false by default, which is
the next commit.

Change-Id: I3c098a9a5055ec8f2b64319c6cadd13c030ba76a