Fix Block::newLoad for IPv6 range blocks
If Block::newLoad finds multiple IP range blocks, the block with the narrowest
range is chosen. In order to determine this, the start and end of the range are
converted to decimal. Prior to this patch, this does not work for IPv6 because
the converter does not handle the 'v6-' prefix on IPv6 addresses, so remove
this.

Bug: T222246
Change-Id: Ie8bebd86d8b4a4d2c7bffd1edaa09ec5ba207303