resourceloader: Minor optimisation in variable declarations
Trying out Safari after a while and, it's Dev Tools have a
way to see the compiler's type inference, showing that Safari
is not able to always detect some numbers as non-optional.

While I generally avoid optimising for an optimiser, I believe
being more explicit is always better or equal. Plus, it saves a
few bytes while at it.

Change-Id: I4084452a74c51953ff0a1609b3b52d696f14a7e1