don't die producing xml files if rev text export conversion fails
In abstracts for the specific case, we don't care at all, since the
problem is that it's a self redirect. Redirects are filtered out of
the stream at the end so it won't even show up.

In anything else, we do what dumpTextPass does already, which is to
leave the text alone and emit it as is.

Bug: T217329
Change-Id: I39cdf89531c67962b1a9bba4e0a91f7c655ad6f3