Add optional jQuery event delegates in Confirmable
Since we want to use the Confirmable on pages with dynamic content
(RecentChanges for example), we need to add support for event
delegation to confirmables, as this is the most reliable way to attach
the event to new elements on the page.

As a side effect, it should also improve performance since it
reduces the amount of event handlers needed for our use case.

Bug: T218354
Change-Id: I22bea39c371d329b40b76ebadc9b74e6d06bfe13