Use mediawiki/mediawiki-phan-config 0.3.0
Use the library instead of duplicating most of the config/defaults that it
provides. MediaWiki core is different of course, so we have to override a
bunch of file/directory lists, but there was a lot being duplicated.

This is the first step in migrating to a newer phan version.

Change-Id: Ib5987ebdf208138d97e1aba8ef54438064063fe9