Upgrade to newer phan
MediaWiki core now runs phan 1.2.6, bringing in nearly 2+ years of upstream
fixes.

Configuration was moved from `tests/phan` to `.phan/`. The legacy bash wrapper
script is still kept in the repository in its own location for any extensions
that are still using it. It should be removed before 1.33 is released.

Since there's a lot of new issues being flagged, all currently failing issues
are suppressed, and will be fixed in follow-up patches.

We're dropping the jetbrains/phpstorm-stubs repository in favor of just
the minimal stubs we need. Stubs for PHP extensions are kept in
the new `.phan/internal_stubs` directory, since they're in a slightly
different format than normal stubs.

Normal stubs are kept in `.phan/stubs`. wikidiff2 and excimer are kept with
these since we're also the upstream for them.

Change-Id: I3fe437befa17f4fbaf97aa6271f659b56021f396