block: Avoid use of empty() which doesn't warn on var non-existence
As empty() is more tolerant and doesn't warn on variable non-existence
and will go ahead and treat it as if it has the value null, it's better
to use ! in this case to make sure we catch even things like typos in
the variable names within the code base.

Let's make sure intended checks are actually checked instead of being
tolerant. Follow up on: Ibaf555fac293da46fa397ccb9, thanks @Krinkle.

Change-Id: Ibc45dba373bbbb6735a9681ed07a2438d4073ee0