resources: Improve foreign-resources docs and error messages
Previously, the help message was only shown when invoking the script
without parameters. Now, if invoking it with incorrect or mispelled
parameters, it shows what you can do instead (and no more exception trace,
which it did before).

Also document clearly but briefly how to use it, and with examples.

Change-Id: Ib5ba3dc02c54039495a33d6e0cc4d026ca708ba1