specialpage: Remove unused QueryPage::doFeed()
This code was written in 2004, and pretty much untouched since then.

The last non-trivial update to this code was done in 2014 via I1de6301.
But as far as I can tell this was just maintenance, not motivated by
somebody actually using this code.

I suspect this is unused for a long time already, long before 2014.
Possibly even unused in 2004. Really: I tried to dig into the revision
history of the core codebase to find when and where this was used, and
when it became unused. I could not find anything. I also used
https://codesearch.wmflabs.org to make sure there are no known callers
or subclasses hidden anywhere. I could not find anything either.

In conclusion I suggest to skip any deprecation phase.

Change-Id: I5f52c44760b903bcb348bebf8c6a0f6faf8180b7