title: Fix bogus Title::getSkinFromConfigSubpage()
Same as I3d8c6f6 already did for two other methods. It looks like this
one was forgotten.

This patch also adds tests for the previously untested method
Title::getSkinFromConfigSubpage().

Note this patch changes the behavior of Title::getSkinFromConfigSubpage()
slightly. It will return an empty string much more often. This is not a
breaking change as far as I can tell. According to
https://codesearch.wmflabs.org/search/?q=getSkinFromConfigSubpage%5C(
there are only two known callers, and both guarantee the string passed to
this method contains both a slash and a dot. Returning an empty string in
all other cases is much closer to the original intent of the method. For
example. "User:Vector" will not be considered a valid skin name any more.

Bug: T229443
Change-Id: I0a5be39ca11a4e5d6c05f70eb05e80c6c4980523