Consistently use classes for deleted items in lists of diffs
Even if a revision has been deleted and the history-deleted class is
being used, the original class for its corresponding element when
not deleted should be applied. This is important as it allows skins
to skin consistently. "history-deleted" is a modifier class - it provides
further information on the original meaning.

This blocks styling the history page from core in Minerva skin

Bug: T216420
Change-Id: Ia659606838d7e38bc09054e36cd980b00f5e6da9