title: Rewrite Title::getBaseText() for performance/readability
This code does not really need to explode a path in all it's individual
segments. All it needs to know is *if* a title is a path (with at least
one slash), and then remove the last segment. That's pretty much exactly
what the code proposed in this patch now does.

Change-Id: I9cf714f7bb637ecd9e1f44340219c76ddd60ee29