resourceloader: Improve test cases for DerivativeResourceLoaderContext
Make the methods more descriptive of the stories and use cases
they cover, and less method-oriented given they don't only
just test a getter or setter anyway (and doing so is arguably
not useful).

Also fold the "testGetInherited" case into the specific ones,
add additional assertions for inheritence that was previously
not tested (specifically, for getVersion and getRaw).

Change-Id: I01c72112e517e66cea5ecf9e2ab53b9039fb99bb