installer: Tweak MysqlUpdater::doWatchlistUpdate query
Instead of doing two near table scans in a race condition prone way,
just figure out the rows to INSERT in a UNION SELECT query. The
query should see a single state of the DB, so intermediate changes
are not such a problem. Previously, they would trigger the large and
slow INSERT SELECT.

Bug: T160993
Change-Id: If6c2ce67a97200f3db752c1edc1da6975f154808