Change job table params from blob to mediumblob
The default blob type holds 65kB which is often enough, but some
jobs push over this limit. When using mysql these are silently
truncated to 65kB which only issues errors when the data is retrieved.

Increase from blob to mediumblob which gives a max size of 16MB. There
may be some valid use cases for jobs > 64kB, but there should be no
use case for jobs > 16MB.

Bug: T124196
Change-Id: I6df6634bd3a5b5fa27ce6f8ff20d251a9b5eba32