Make the ratio of requests getting Priority Hints configurable
This will allow to create 2 separate populations in
order to verify the effect of the hint.

It's fine if it's not exactly balanced in practice,
what matters is getting a significant amount of
traffic for both scenarios (origin trial is enabled +
an image is given high prio/origin trial is enabled +
no image is given special treatment).

Bug: T216499
Change-Id: I373960b2bed8437c2e97e6d729d43aff6901046c