Fix PhanPluginDuplicateExpressionBinaryOp in DjVuImage (#9)
== has a higher precedence operator than &.

So the line `$chunkLength & 1 == 1` is interepreted as:
`$chunkLength & true`. Probably not what was intended, but it actually
works out because true is cast to 1, and the expression will either
result with 0 (falsey) or 1 (truthy), and we were looking for it to
equal 1 anyways. Incredible.

I assume that it was supposed to be `( $chunkLength & 1 ) == 1`, so I've
modified it to use that.

Change-Id: If03823b9286e0d8282519949bf46e2ba4cca0843