Fix phan errors by improving documentation (#10)
PhanTypeExpectedObjectPropAccess was flagged by phan in the
DatabaseMysqli and Preprocessor_Hash classes.

In Database, the $conn property might be a standard object, such as
`\mysqli`, which is not a resource.

In Preprocessor, phan was getting confused and thinking
PPDStack::getCurrentPart() was returning bool, and not a PPDPart object.
Adding explicit documentation about the return value fixed that.

Change-Id: I0a3aa219693da5cb46ff9c0936841ed740c6968a