resourceloader: Add missing (string) cast to ClientHtml tests
One of the assertions already did this, but the others not yet.

The reason we do this is because the return value is a WrappedString
object. The assertEquals() method is able to satisfy the expected
value because of loose comparison casting it to a string for the
case where it matches.

But, when it fails during development, PHPUnit was still printing
the object as "Actual" and the string as "Expected" which makes
it difficult to understand what's going on, and also no diff.

Change-Id: Ice9fe5061fb76867e9bee2f34c88d3ddac3144f9